Move cluster query processing into separate module
