+++ cli/src/lib.rs
@@ -2,6 +2,7 @@
 extern crate lazy_static;
 
 pub mod cli;
+pub mod cluster_query;
 pub mod config;
 pub mod display;
 pub mod input_parsers;
